requirements.txt
requests==2.31.0
python-dotenv==0.21.1
Flask==3.0.2
Flask-Bootstrap==3.3.7.1 